+++ test/svlog/regr/iss190_assign_expr.sv
@@ -1,7 +1,20 @@
-// RUN: moore %s -e foo -e bar
-// IGNORE  see #190
+// RUN: moore %s -e foo
 
 module foo;
-	int i;
-	initial for (i = 1; i < 16; i = i + 1);
+    int i, j;
+    initial begin
+        i = (j = 1);
+        i = (j += 1);
+        i = (j -= 1);
+        i = (j *= 1);
+        i = (j /= 1);
+        i = (j %= 1);
+        i = (j &= 1);
+        i = (j |= 1);
+        i = (j ^= 1);
+        i = (j <<= 1);
+        i = (j >>= 1);
+        i = (j <<<= 1);
+        i = (j >>>= 1);
+    end
 endmodule